+++ packages/health/ios/Classes/SwiftHealthPlugin.swift
@@ -21,7 +21,12 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
     let BODY_FAT_PERCENTAGE = "BODY_FAT_PERCENTAGE"
     let BODY_MASS_INDEX = "BODY_MASS_INDEX"
     let BODY_TEMPERATURE = "BODY_TEMPERATURE"
+    let DIETARY_CARBS_CONSUMED = "DIETARY_CARBS_CONSUMED"
+    let DIETARY_ENERGY_CONSUMED = "DIETARY_ENERGY_CONSUMED"
+    let DIETARY_FATS_CONSUMED = "DIETARY_FATS_CONSUMED"
+    let DIETARY_PROTEIN_CONSUMED = "DIETARY_PROTEIN_CONSUMED"
     let ELECTRODERMAL_ACTIVITY = "ELECTRODERMAL_ACTIVITY"
+    let FORCED_EXPIRATORY_VOLUME = "FORCED_EXPIRATORY_VOLUME"
     let HEART_RATE = "HEART_RATE"
     let HEART_RATE_VARIABILITY_SDNN = "HEART_RATE_VARIABILITY_SDNN"
     let HEIGHT = "HEIGHT"
@@ -40,6 +45,8 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
     let SLEEP_IN_BED = "SLEEP_IN_BED"
     let SLEEP_ASLEEP = "SLEEP_ASLEEP"
     let SLEEP_AWAKE = "SLEEP_AWAKE"
+    let EXERCISE_TIME = "EXERCISE_TIME"
+    let WORKOUT = "WORKOUT"
 
 
     public static func register(with registrar: FlutterPluginRegistrar) {
@@ -65,6 +72,15 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
         else if (call.method.elementsEqual("getData")){
             getData(call: call, result: result)
         }
+
+        /// Handle writeData
+        else if (call.method.elementsEqual("writeData")){
+            writeData(call: call, result: result)
+        }
+
+        else if (call.method.elementsEqual("getStepsData")){
+            getStepsData(call: call, result: result)
+        }
     }
 
     func checkIfHealthDataAvailable(call: FlutterMethodCall, result: @escaping FlutterResult) {
@@ -83,13 +99,43 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
         }
 
         if #available(iOS 11.0, *) {
-            healthStore.requestAuthorization(toShare: nil, read: typesToRequest) { (success, error) in
+            healthStore.requestAuthorization(toShare: typesToRequest, read: typesToRequest) { (success, error) in
                 result(success)
             }
-        } 
+        }
         else {
             result(false)// Handle the error here.
         }
+    }
+
+    func writeData(call: FlutterMethodCall, result: @escaping FlutterResult){
+        guard let arguments = call.arguments as? NSDictionary,
+            let value = (arguments["value"] as? Double),
+            let type = (arguments["dataTypeKey"] as? String),
+            let startDate = (arguments["startTime"] as? NSNumber),
+            let endDate = (arguments["endTime"] as? NSNumber)
+            else {
+                print("writeData: Invalid arguments")
+                return
+            }
+
+        let dateFrom = Date(timeIntervalSince1970: startDate.doubleValue / 1000)
+        let dateTo = Date(timeIntervalSince1970: endDate.doubleValue / 1000)
+
+        print("Successfully called writeData with value of \(value) and type of \(type)")
+
+        let quantity = HKQuantity(unit: unitLookUp(key: type), doubleValue: value)
+
+        let sample = HKQuantitySample(type: dataTypeLookUp(key: type) as! HKQuantityType, quantity: quantity, start: dateFrom, end: dateTo)
+
+        HKHealthStore().save(sample, withCompletion: { (success, error) in
+            if let err = error {
+                print("Error Saving \(type) Sample: \(err.localizedDescription)")
+            }
+            DispatchQueue.main.async {
+                result(success)
+            }
+        })
     }
 
     func getData(call: FlutterMethodCall, result: @escaping FlutterResult) {
@@ -97,6 +143,7 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
         let dataTypeKey = (arguments?["dataTypeKey"] as? String) ?? "DEFAULT"
         let startDate = (arguments?["startDate"] as? NSNumber) ?? 0
         let endDate = (arguments?["endDate"] as? NSNumber) ?? 0
+        let limit = (arguments?["limit"] as? Int) ?? HKObjectQueryNoLimit
 
         // Convert dates from milliseconds to Date()
         let dateFrom = Date(timeIntervalSince1970: startDate.doubleValue / 1000)
@@ -104,17 +151,30 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
 
         let dataType = dataTypeLookUp(key: dataTypeKey)
         let predicate = HKQuery.predicateForSamples(withStart: dateFrom, end: dateTo, options: .strictStartDate)
-        let sortDescriptor = NSSortDescriptor(key: HKSampleSortIdentifierEndDate, ascending: true)
+        let sortDescriptor = NSSortDescriptor(key: HKSampleSortIdentifierEndDate, ascending: false)
 
-        let query = HKSampleQuery(sampleType: dataType, predicate: predicate, limit: HKObjectQueryNoLimit, sortDescriptors: [sortDescriptor]) {
+        let query = HKSampleQuery(sampleType: dataType, predicate: predicate, limit: limit, sortDescriptors: [sortDescriptor]) {
             x, samplesOrNil, error in
 
-            guard let samples = samplesOrNil as? [HKQuantitySample] else {
-                guard var samplesCategory = samplesOrNil as? [HKCategorySample] else {
-                    result(FlutterError(code: "FlutterHealth", message: "Results are null", details: "\(error)"))
-                    return
+            switch samplesOrNil {
+            case let (samples as [HKQuantitySample]) as Any:
+
+                DispatchQueue.main.async {
+                    result(samples.map { sample -> NSDictionary in
+                        let unit = self.unitLookUp(key: dataTypeKey)
+
+                        return [
+                            "uuid": "\(sample.uuid)",
+                            "value": sample.quantity.doubleValue(for: unit),
+                            "date_from": Int(sample.startDate.timeIntervalSince1970 * 1000),
+                            "date_to": Int(sample.endDate.timeIntervalSince1970 * 1000),
+                            "source_id": sample.sourceRevision.source.bundleIdentifier,
+                            "source_name": sample.sourceRevision.source.name
+                        ]
+                    })
                 }
-                print(samplesCategory)
+
+            case var (samplesCategory as [HKCategorySample]) as Any:
                 if (dataTypeKey == self.SLEEP_IN_BED) {
                     samplesCategory = samplesCategory.filter { $0.value == 0 }
                 }
@@ -124,34 +184,39 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
                 if (dataTypeKey == self.SLEEP_ASLEEP) {
                     samplesCategory = samplesCategory.filter { $0.value == 1 }
                 }
-                result(samplesCategory.map { sample -> NSDictionary in
-                    let unit = self.unitLookUp(key: dataTypeKey)
 
-                    return [
-                        "uuid": "\(sample.uuid)",
-                        "value": sample.value,
-                        "date_from": Int(sample.startDate.timeIntervalSince1970 * 1000),
-                        "date_to": Int(sample.endDate.timeIntervalSince1970 * 1000),
-                        "source_id": sample.sourceRevision.source.bundleIdentifier,
-                        "source_name": sample.sourceRevision.source.name
-                    ]
-                })
+                DispatchQueue.main.async {
+                    result(samplesCategory.map { sample -> NSDictionary in
+                        return [
+                            "uuid": "\(sample.uuid)",
+                            "value": sample.value,
+                            "date_from": Int(sample.startDate.timeIntervalSince1970 * 1000),
+                            "date_to": Int(sample.endDate.timeIntervalSince1970 * 1000),
+                            "source_id": sample.sourceRevision.source.bundleIdentifier,
+                            "source_name": sample.sourceRevision.source.name
+                        ]
+                    })
+                }
+
+            case let (samplesWorkout as [HKWorkout]) as Any:
+                DispatchQueue.main.async {
+                    result(samplesWorkout.map { sample -> NSDictionary in
+                        return [
+                            "uuid": "\(sample.uuid)",
+                            "value": Int(sample.duration),
+                            "date_from": Int(sample.startDate.timeIntervalSince1970 * 1000),
+                            "date_to": Int(sample.endDate.timeIntervalSince1970 * 1000),
+                            "source_id": sample.sourceRevision.source.bundleIdentifier,
+                            "source_name": sample.sourceRevision.source.name
+                        ]
+                    })
+                }
+
+            default:
                 return
             }
-            result(samples.map { sample -> NSDictionary in
-                let unit = self.unitLookUp(key: dataTypeKey)
-
-                return [
-                    "uuid": "\(sample.uuid)",
-                    "value": sample.quantity.doubleValue(for: unit),
-                    "date_from": Int(sample.startDate.timeIntervalSince1970 * 1000),
-                    "date_to": Int(sample.endDate.timeIntervalSince1970 * 1000),
-                    "source_id": sample.sourceRevision.source.bundleIdentifier,
-                    "source_name": sample.sourceRevision.source.name
-                ]
-            })
-            return
         }
+
         HKHealthStore().execute(query)
     }
 
@@ -178,16 +243,21 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
         unitDict[BLOOD_PRESSURE_SYSTOLIC] = HKUnit.millimeterOfMercury()
         unitDict[BODY_FAT_PERCENTAGE] = HKUnit.percent()
         unitDict[BODY_MASS_INDEX] = HKUnit.init(from: "")
-        unitDict[BODY_TEMPERATURE] = HKUnit.degreeCelsius()
+        unitDict[BODY_TEMPERATURE] = HKUnit.degreeFahrenheit()
+        unitDict[DIETARY_CARBS_CONSUMED] = HKUnit.gram()
+        unitDict[DIETARY_ENERGY_CONSUMED] = HKUnit.kilocalorie()
+        unitDict[DIETARY_FATS_CONSUMED] = HKUnit.gram()
+        unitDict[DIETARY_PROTEIN_CONSUMED] = HKUnit.gram()
         unitDict[ELECTRODERMAL_ACTIVITY] = HKUnit.siemen()
+        unitDict[FORCED_EXPIRATORY_VOLUME] = HKUnit.liter()
         unitDict[HEART_RATE] = HKUnit.init(from: "count/min")
         unitDict[HEART_RATE_VARIABILITY_SDNN] = HKUnit.secondUnit(with: .milli)
-        unitDict[HEIGHT] = HKUnit.meter()
+        unitDict[HEIGHT] = HKUnit.inch()
         unitDict[RESTING_HEART_RATE] = HKUnit.init(from: "count/min")
         unitDict[STEPS] = HKUnit.count()
         unitDict[WAIST_CIRCUMFERENCE] = HKUnit.meter()
         unitDict[WALKING_HEART_RATE] = HKUnit.init(from: "count/min")
-        unitDict[WEIGHT] = HKUnit.gramUnit(with: .kilo)
+        unitDict[WEIGHT] = HKUnit.pound()
         unitDict[DISTANCE_WALKING_RUNNING] = HKUnit.meter()
         unitDict[FLIGHTS_CLIMBED] = HKUnit.count()
         unitDict[WATER] = HKUnit.liter()
@@ -195,9 +265,11 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
         unitDict[SLEEP_IN_BED] = HKUnit.init(from: "")
         unitDict[SLEEP_ASLEEP] = HKUnit.init(from: "")
         unitDict[SLEEP_AWAKE] = HKUnit.init(from: "")
+        unitDict[EXERCISE_TIME] =  HKUnit.minute()
+        unitDict[WORKOUT] = HKUnit.init(from: "")
 
         // Set up iOS 11 specific types (ordinary health data types)
-        if #available(iOS 11.0, *) { 
+        if #available(iOS 11.0, *) {
             dataTypesDict[ACTIVE_ENERGY_BURNED] = HKSampleType.quantityType(forIdentifier: .activeEnergyBurned)!
             dataTypesDict[BASAL_ENERGY_BURNED] = HKSampleType.quantityType(forIdentifier: .basalEnergyBurned)!
             dataTypesDict[BLOOD_GLUCOSE] = HKSampleType.quantityType(forIdentifier: .bloodGlucose)!
@@ -207,7 +279,12 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
             dataTypesDict[BODY_FAT_PERCENTAGE] = HKSampleType.quantityType(forIdentifier: .bodyFatPercentage)!
             dataTypesDict[BODY_MASS_INDEX] = HKSampleType.quantityType(forIdentifier: .bodyMassIndex)!
             dataTypesDict[BODY_TEMPERATURE] = HKSampleType.quantityType(forIdentifier: .bodyTemperature)!
+            dataTypesDict[DIETARY_CARBS_CONSUMED] = HKSampleType.quantityType(forIdentifier: .dietaryCarbohydrates)!
+            dataTypesDict[DIETARY_ENERGY_CONSUMED] = HKSampleType.quantityType(forIdentifier: .dietaryEnergyConsumed)!
+            dataTypesDict[DIETARY_FATS_CONSUMED] = HKSampleType.quantityType(forIdentifier: .dietaryFatTotal)!
+            dataTypesDict[DIETARY_PROTEIN_CONSUMED] = HKSampleType.quantityType(forIdentifier: .dietaryProtein)!
             dataTypesDict[ELECTRODERMAL_ACTIVITY] = HKSampleType.quantityType(forIdentifier: .electrodermalActivity)!
+            dataTypesDict[FORCED_EXPIRATORY_VOLUME] = HKSampleType.quantityType(forIdentifier: .forcedExpiratoryVolume1)!
             dataTypesDict[HEART_RATE] = HKSampleType.quantityType(forIdentifier: .heartRate)!
             dataTypesDict[HEART_RATE_VARIABILITY_SDNN] = HKSampleType.quantityType(forIdentifier: .heartRateVariabilitySDNN)!
             dataTypesDict[HEIGHT] = HKSampleType.quantityType(forIdentifier: .height)!
@@ -223,6 +300,8 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
             dataTypesDict[SLEEP_IN_BED] = HKSampleType.categoryType(forIdentifier: .sleepAnalysis)!
             dataTypesDict[SLEEP_ASLEEP] = HKSampleType.categoryType(forIdentifier: .sleepAnalysis)!
             dataTypesDict[SLEEP_AWAKE] = HKSampleType.categoryType(forIdentifier: .sleepAnalysis)!
+            dataTypesDict[EXERCISE_TIME] = HKSampleType.quantityType(forIdentifier: .appleExerciseTime)!
+            dataTypesDict[WORKOUT] = HKSampleType.workoutType()
 
             healthDataTypes = Array(dataTypesDict.values)
         }
@@ -242,8 +321,44 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
         // Concatenate heart events and health data types (both may be empty)
         allDataTypes = Set(heartRateEventTypes + healthDataTypes)
     }
+
+    private func getStepsData(call: FlutterMethodCall, result: @escaping FlutterResult) {
+        let arguments = call.arguments as? NSDictionary
+        let startDate = (arguments?["startDate"] as? NSNumber) ?? 0
+        let endDate = (arguments?["endDate"] as? NSNumber) ?? 0
+        let dateFrom = Date(timeIntervalSince1970: startDate.doubleValue / 1000)
+        let dateTo = Date(timeIntervalSince1970: endDate.doubleValue / 1000)
+
+        let predicate = HKQuery.predicateForSamples(withStart: dateFrom,
+                                                    end: dateTo,
+                                                    options: [])
+        let calendar = Calendar.current
+        let anchorDate = calendar.date(bySettingHour: 0, minute: 0, second: 0, of: Date())
+        let query = HKStatisticsCollectionQuery(quantityType: HKObjectType.quantityType(forIdentifier: .stepCount)!,
+                                                quantitySamplePredicate: predicate,
+                                                options: .cumulativeSum,
+                                                anchorDate: anchorDate!,
+                                                intervalComponents: DateComponents(day: 1))
+
+        var data: [NSDictionary] = []
+        query.initialResultsHandler = { _, results, _ in
+            guard let statsCollection = results else { return }
+            statsCollection.enumerateStatistics(from: dateFrom, to: dateTo) { statistics, _ in
+                if let quantity = statistics.sumQuantity() {
+                    let stepValue = floor(quantity.doubleValue(for: HKUnit.count()))
+                    data.append( [
+                        "uuid": "",
+                        "source_id": "",
+                        "source_name": "",
+                        "date_from": Int(statistics.startDate.timeIntervalSince1970 * 1000),
+                        "date_to": Int(statistics.endDate.timeIntervalSince1970 * 1000),
+                        "value": stepValue,
+                    ])
+                }
+            }
+            result(data)
+            return
+        }
+        HKHealthStore().execute(query)
+    }
 }
-
-
-
-