+++ packages/health/lib/src/health_factory.dart
@@ -91,10 +91,43 @@ class HealthFactory {
     return bmiHealthPoints;
   }
 
-  /// Get an list of [HealthDataPoint] from an list of [HealthDataType].
+  ///
+  /// Saves health data into the HealthKit or Google Fit store
+  ///
+  /// Returns a Future of true if successful, a Future of false otherwise
+  ///
+  /// Parameters
+  ///
+  /// [value]
+  ///   value of the health data in double
+  /// [type]
+  ///   the value's HealthDataType
+  /// [startTime]
+  ///   a DateTime object that specifies the start time when this data value is measured.
+  ///   It must be equal to or earlier than [endTime]
+  /// [endTime]
+  ///   a DateTime object that specifies the end time when this value is measured.
+  ///   It must be equal to or later than [startTime].
+  ///   Simply set [endTime] equal to [startTime]
+  ///   if the value is measured only at a specific point in time.
+  ///
+  Future<bool> writeHealthData(double value, HealthDataType type,
+      DateTime startTime, DateTime endTime) async {
+    if (startTime.isAfter(endTime))
+      throw ArgumentError("startTime must be equal or earlier than endTime");
+    Map<String, dynamic> args = {
+      'value': value,
+      'dataTypeKey': _enumToString(type),
+      'startTime': startTime.millisecondsSinceEpoch,
+      'endTime': endTime.millisecondsSinceEpoch
+    };
+    bool? success = await _channel.invokeMethod('writeData', args);
+    return success ?? false;
+  }
+
   Future<List<HealthDataPoint>> getHealthDataFromTypes(
       DateTime startDate, DateTime endDate, List<HealthDataType> types) async {
-    final dataPoints = <HealthDataPoint>[];
+    List<HealthDataPoint> dataPoints = [];
 
     for (var type in types) {
       final result = await _prepareQuery(startDate, endDate, type);
@@ -180,5 +213,49 @@ class HealthFactory {
       }
     }
     return unique;
+  }
+
+  Future<List<HealthDataPoint>> getStepData(
+      DateTime startDate, DateTime endDate) async {
+    final dataType = HealthDataType.STEPS;
+    if (_platformType != PlatformType.IOS) {
+      throw _HealthException(dataType, 'getStepData is only iOS');
+    }
+
+    _deviceId ??= (await _deviceInfo.iosInfo).identifierForVendor;
+
+    final dataPoints = <HealthDataPoint>[];
+    final args = <String, dynamic>{
+      'dataTypeKey': _enumToString(dataType),
+      'startDate': startDate.millisecondsSinceEpoch,
+      'endDate': endDate.millisecondsSinceEpoch
+    };
+
+    final unit = _dataTypeToUnit[dataType]!;
+
+    final fetchedDataPoints = await _channel.invokeMethod('getStepsData', args);
+    if (fetchedDataPoints != null) {
+      return fetchedDataPoints.map<HealthDataPoint>((e) {
+        final num value = e['value'];
+        final DateTime from =
+            DateTime.fromMillisecondsSinceEpoch(e['date_from']);
+        final DateTime to = DateTime.fromMillisecondsSinceEpoch(e['date_to']);
+        final String sourceId = e["source_id"];
+        final String sourceName = e["source_name"];
+        return HealthDataPoint(
+          value,
+          dataType,
+          unit,
+          from,
+          to,
+          _platformType,
+          _deviceId!,
+          sourceId,
+          sourceName,
+        );
+      }).toList();
+    } else {
+      return <HealthDataPoint>[];
+    }
   }
 }